refactor(ui): move the Home start command into its own file
screen_home.go was the only production file over the 200-line ceiling. The
start command is a distinct concern from the model and its update loop, and
the split mirrors screen_typing_actions.go.

Also rename the app package's polish test file, which carried a plan phase
number rather than describing what it covers: degraded-size rendering and the
quit prompt.
