+++ internal/app/degraded_and_quit_prompt_test.go
